Replace github-markdown with commonmarker
The github-markdown gem is no longer maintained. To make things work
with commonmarker, its version has to be fixed at 0.23, since
github-markup expectes commonmarker to provide the CommonMarker module,
whereas commonmarker 1.0 provides Commonmarker instead.

There is a small change in the number of newlines output in the
resulting HTML, but this is not a structural change.